rel-html: add support for folder stable lookup
Its not convenient to have to always have to add
a new directory for a release and if you want to minimize
the number of files in a directory we should be able to
split up releases into a stable/ directory and for each
release have in it the tarballs alone for that target
release. This adds initial support for that.

For example if you have a stable/ directory and the
current testing release is v3.7-rc1 it will also look
under stable/v3.7-rc1/ for release tarballs there.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>