rel-html: update compat-drivers release values
This updates the compat-drivers config file to account for
the latest releases. This type of change is one that we
should strive to automate but its unclear exactly how to
best do this for any random project. One possible idea
is to add support into git information about the latest
releases, or to have projects list this somehow on github
and use github API to extract the release information.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>