rel-html: document how to contribute to this project
This relies on the new signed-off-by project which
tries to document the Developer Certificate of Origin
process used by the Linux kernel and allows other
projects to easily document following such same procedures.

In short:

This project tracks patch provenance and licensing using the Developer
Certificate of Origin and Signed-off-by tags initially developed by
the Linux kernel project.  Because the documentation for this
procedure is licensed under the GPLv2, we have chosen not to include
it in our project directly.  Instead, please see:

http://git.tremily.us/?p=signed-off-by.git;a=blob;f=Documentation/SubmittingPatches;h=34055986ab836553896f091225448c448a4cc62c;hb=refs/heads/signed-off-by

For details see:

https://github.com/wking/signed-off-by/
https://lkml.org/lkml/2012/11/20/636

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>