rel-html: fix release mod signature checks
The signed files for releases should be the tarballs
and not the compressed files themselves. This follows
the kernel.org release model by which a file uploader
is used to upload the tarball alone and the compression
is later done by the file uploader on the server side
once the tarball alone as been verified.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>