rel-html: infer releases
This adds support to infer releases based on the
major and minor numbers of a release (aka, VERSION,
PATCHLEVEL). This removes the requirements for updating
the configuration file every time a new release is made.
An update to the configuration file is only needed now
when its decided a new release is EOL'd or a new major/minor
number gets added to the release model.

So for example for the Linux kernel we make these changes
to the configuration file for it:

	-rel_html_stable_vers   = 3.7.2
	-                         3.6.11:EOL
	-                         3.5.7:EOL
	-                         3.4.25
	-                         3.3.8:EOL
	-                         3.2.36
	-                         3.0.58
	-                         2.6.34.13
	-                         2.6.32.60
	-rel_html_testing_ver   = 3.8-rc3
	+supported =
	+               3.8
	+               3.7
	+               3.4
	+               3.2
	+               3.0
	+               2.6.34
	+               2.6.32
	+eol =
	+               3.6
	+               3.5
	+               3.3

This is what a releas page looks like for it:

http://drvbp1.linux-foundation.org/~mcgrof/rel-html/linux/

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>