rel-html: fix next release mode updates
The linux-next based releases can have mod updates,
so for example for compat-drivers we can have
compat-drivers-2013-03-07 and an assumption was
made we'd always have a -1 release, followed by a
possible -2 release. In practice we've only been
making one release per day though so the mod update
is blank. Fix this by making the regexp that captures
the release mode update as a digit. While at it update
the stable release regexps too to use digit for the
regexp.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>