rel-html: fix stable release attribute search
Each release has a set of things we need to verify
prior to publishing a URL for it. For example a stable
release requires:

  * ChangeLog
  * Signed ChangeLog
  * Tarball
  * Signed tarball

We search for these files based on a pivot 'release'
version string. The existing algorithm was using the
full release version but given that for example
compat-drivers relies on only one ChangeLog for
the different modified release types the search was
skipping out on finding a ChangeLog for modified
stable releases.

So for example compat-drivers-3.8.2-2-su relase uses
the ChangeLog-3.8.2-2, but as it was before this commit
the search was looking for "3.8.2-2-su" in the URL
attributes it was looking, not 3.8.2. This change
modifies parsing the URLs it looks at by looking at
the short version instead.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>