rel-html: rename compat-drivers.cfg to backports.cfg
The compat-drivers project was renamed to backports.
Rename the project file and also update the config file
to add the v3.10 kernel as supported.

We also add the ignore_changelogs for this project given
that we want to skip these moving forward and simply provide
shortlogs upon releases and refer folks to use git for a full
changelog.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>