rel-html: really listen to ignore_changelogs
When a project was setting ignore_changelogs the changelog
was still printed if and only if the changelog happened
to be found by chance and it was also signed. In practice
if a project wants to abandon changelogs this would then
produce an output that would likely make users think the
developers forgot to provide a changelog. Simply ignore
changelogs even if they are found for older releases when
ignore_changelogs is set.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>