rel-html: add RFC3339 / short style next release types support
A project might be using RFC3339 (2013-05-02) style release
names for their latest releases and if they switch to the
smaller style (20130502) we need to treat both release
styles as possible. Add support for this.

Your project need not do anything if you switch, the tarball
hunter will everything for for you.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>