backports: eol the same Linux releases
It just so happens that the same releases on Linux that are EOL'd
are also the old name releases for backports, ie. 'compat-drivers'.
These are now deprecated and you should be using the release marked
as supported. These should match the latest supported stable
kernel releases.

Note that for backports you should be using the *latest* stable
release. The release will work on *older* kernel releases. That
is the point.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>