rel-html: revise stable releases prior to display
rel-html currently will look for each supported release tag
on earch base URL it enters and if it finds any major release
it'll eventually display it. This is displaying some old
stable releases on supported releases which have sublevels
on the releases. This adds a sanity check on the releases
that revises the weights of each release against the supplied
release tags.

The final stable releases shown are only the releases that
have the *max weight* for the supplied supported release tags.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>